Fix BUILD_SHARED_LIBS=ON
MLIR currently has two issues with building shared libraries:
1) There are circular dependencies between libraries
2) The registration of Dialects is handled with static initializers

The simplest solution to the above is to force all of the MLIR libraries
to be STATIC, even when building most of LLVM as shared libraries.  This
isn't too bad as a short-term solution, since MLIR is small enough that
it links pretty quickly.